Delete  article confirmation behaviour updated

+++ app/src/main/java/com/mrwhoknows/csgeeks/ui/admin_page/articles/YourArticlesFragment.kt
@@ -3,16 +3,13 @@ package com.mrwhoknows.csgeeks.ui.admin_page.articles
 import android.os.Bundle
 import android.util.Log
 import android.view.View
-import android.widget.RadioButton
-import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.chip.Chip
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.snackbar.Snackbar
 import com.mrwhoknows.csgeeks.R
 import com.mrwhoknows.csgeeks.adapter.ArticleListAdapter
@@ -22,21 +19,16 @@ import com.mrwhoknows.csgeeks.util.LoginInfo
 import com.mrwhoknows.csgeeks.util.Resource
 import com.mrwhoknows.csgeeks.util.Util
 import com.mrwhoknows.csgeeks.viewmodels.BlogViewModel
-import kotlinx.android.synthetic.main.fragment_articles_list.*
-import kotlinx.android.synthetic.main.sort_menu_bottom_sheet.*
+import kotlinx.android.synthetic.main.fragment_manage_articles.*
 
 private const val TAG = "YourArticlesFragment"
 
-class YourArticlesFragment : Fragment(R.layout.fragment_articles_list) {
+class YourArticlesFragment : Fragment(R.layout.fragment_manage_articles) {
 
     private lateinit var articleAdapter: ArticleListAdapter
     private lateinit var viewModel: BlogViewModel
     private var authorName: String = ""
     private var loginToken: String = ""
-    private lateinit var bottomSheetBehavior: BottomSheetBehavior<ConstraintLayout>
-    private var selectedTag: String = ""
-    private var sortBy: String = ""
-    private var order: String = ""
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -52,161 +44,15 @@ class YourArticlesFragment : Fragment(R.layout.fragment_articles_list) {
         viewModel = (activity as AdminActivity).viewModel
 
         showYourArticles()
+        Snackbar.make(
+            requireView(),
+            "Slide article right to delete.",
+            Snackbar.LENGTH_SHORT
+        )
+            .setBackgroundTint(ContextCompat.getColor(requireContext(), R.color.colorNordBlue))
+            .show()
+
         swipeToDeleteArticle()
-        blogTags()
-        sortSheet()
-    }
-
-    private fun sortSheet() {
-        val bottomSheet = sortOptionsBottomSheetLayout
-        bottomSheetBehavior = BottomSheetBehavior.from(bottomSheet)
-        bottomSheetBehavior.state = BottomSheetBehavior.STATE_HIDDEN
-
-        btnSort.setOnClickListener {
-            bottomSheetBehavior.state = BottomSheetBehavior.STATE_EXPANDED
-
-            tvApplyBtn.setOnClickListener {
-                val id = sortByOptions.checkedRadioButtonId
-                val option = requireView().findViewById<RadioButton>(id)
-
-                if (selectedTag.isEmpty()) {
-                    when (option.tag) {
-                        "1" -> {
-                            sortBy = "created"
-                            order = "desc"
-                            viewModel.orderArticlesBy(sortBy, order)
-                        }
-                        "2" -> {
-                            sortBy = "created"
-                            order = "asc"
-                            viewModel.orderArticlesBy(sortBy, order)
-                        }
-                        "3" -> {
-                            sortBy = "title"
-                            order = "asc"
-                            viewModel.orderArticlesBy(sortBy, order)
-                        }
-                        "4" -> {
-                            sortBy = "title"
-                            order = "desc"
-                            viewModel.orderArticlesBy(sortBy, order)
-                        }
-                        "5" -> {
-                            sortBy = "author"
-                            order = "asc"
-                            viewModel.orderArticlesBy(sortBy, order)
-                        }
-                        "6" -> {
-                            sortBy = "author"
-                            order = "desc"
-                            viewModel.orderArticlesBy(sortBy, order)
-                        }
-
-                        else -> viewModel.getAllArticles()
-                    }
-                } else {
-                    when (option.tag) {
-                        "1" -> {
-                            sortBy = "created"
-                            order = "desc"
-                            viewModel.orderArticlesBy(selectedTag, sortBy, order)
-                        }
-                        "2" -> {
-                            sortBy = "created"
-                            order = "asc"
-                            viewModel.orderArticlesBy(selectedTag, sortBy, order)
-                        }
-                        "3" -> {
-                            sortBy = "title"
-                            order = "asc"
-                            viewModel.orderArticlesBy(selectedTag, sortBy, order)
-                        }
-                        "4" -> {
-                            sortBy = "title"
-                            order = "desc"
-                            viewModel.orderArticlesBy(selectedTag, sortBy, order)
-                        }
-                        "5" -> {
-                            sortBy = "author"
-                            order = "asc"
-                            viewModel.orderArticlesBy(selectedTag, sortBy, order)
-                        }
-                        "6" -> {
-                            sortBy = "author"
-                            order = "desc"
-                            viewModel.orderArticlesBy(selectedTag, sortBy, order)
-                        }
-
-                        else -> viewModel.getAllArticles()
-                    }
-                }
-                bottomSheetBehavior.state = BottomSheetBehavior.STATE_HIDDEN
-            }
-        }
-    }
-
-
-    private fun blogTags() {
-        var tags: List<String>
-        viewModel.getArticleTags()
-        viewModel.tags.observe(viewLifecycleOwner, { resource ->
-            when (resource) {
-                is Resource.Success -> {
-                    Util.isLoading(bounceLoader, false)
-                    Util.isLoading(bounceLoaderBG, false)
-                    Log.d(TAG, "tags: ${resource.data}")
-                    resource.data?.let {
-                        tags = it.tags
-                        for (tag in tags) {
-                            val chip = Chip(chipsCategories.context)
-                            chip.text = tag
-                            chip.chipBackgroundColor = ContextCompat.getColorStateList(
-                                requireContext(),
-                                R.color.colorBackgroundDark3
-                            )
-
-                            chip.isClickable = true
-                            chip.isCheckable = true
-                            chipsCategories.addView(chip)
-                        }
-                    }
-                    chipsCategories.isSingleSelection = true
-
-                    chipsCategories.setOnCheckedChangeListener { _, id ->
-                        val chip = chipsCategories.findViewById<Chip>(id)
-                        if (chip != null) {
-                            Log.d(TAG, "chip sel: ${chip.text}")
-                            selectedTag = chip.text.toString()
-                            if (sortBy.isNotEmpty() and order.isNotEmpty())
-                                viewModel.orderArticlesBy(selectedTag, sortBy, order)
-                            else {
-                                sortBy = ""
-                                viewModel.getAllArticles()
-                            }
-                        } else {
-                            Log.d(TAG, "chip not selected")
-                            if (sortBy.isNotEmpty() and order.isNotEmpty())
-                                viewModel.orderArticlesBy(sortBy, order)
-                            else {
-                                selectedTag = ""
-                                viewModel.getAllArticles()
-                            }
-                        }
-                    }
-                }
-                is Resource.Error -> {
-                    Util.isLoading(bounceLoader, true)
-                    Util.isLoading(bounceLoaderBG, true)
-                    Snackbar.make(requireView(), "Something Went Wrong", Snackbar.LENGTH_LONG)
-                        .show()
-                }
-
-                is Resource.Loading -> {
-                    Util.isLoading(bounceLoader, true)
-                    Util.isLoading(bounceLoaderBG, true)
-                }
-            }
-        })
     }
 
     private fun initRecyclerView(data: ArticleList) {
@@ -281,12 +127,23 @@ class YourArticlesFragment : Fragment(R.layout.fragment_articles_list) {
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
                 val pos = viewHolder.adapterPosition
                 val article = articleAdapter.articleMetaList.articles[pos]
-                Snackbar.make(requireView(), "Are you sure?", Snackbar.LENGTH_LONG).apply {
-                    setAction("YES") {
+
+                MaterialAlertDialogBuilder(requireContext())
+                    .setTitle("Do you want to delete article?")
+                    .setMessage("Are you sure?")
+                    .setBackground(
+                        ContextCompat.getDrawable(
+                            requireContext(),
+                            R.color.colorBackgroundDark3
+                        )
+                    )
+                    .setNegativeButton("No") { _, _ ->
+                        articleAdapter.notifyDataSetChanged()
+                    }
+                    .setPositiveButton("Yes") { _, _ ->
                         deleteArticle(article)
                     }
-                    show()
-                }
+                    .show()
             }
         }
 
@@ -297,12 +154,11 @@ class YourArticlesFragment : Fragment(R.layout.fragment_articles_list) {
 
     private fun deleteArticle(article: ArticleList.Article) {
         viewModel.deleteArticleToServer(article.id.toString(), loginToken)
-        viewModel.deleteArticleResponse.observe(viewLifecycleOwner, {
-            when (it) {
+        viewModel.deleteArticleResponse.observe(viewLifecycleOwner, { selectedArticle ->
+            when (selectedArticle) {
                 is Resource.Success -> {
                     Util.isLoading(bounceLoader, false)
                     Util.isLoading(bounceLoaderBG, false)
-                    showYourArticles()
                 }
                 is Resource.Loading -> {
                     Util.isLoading(bounceLoader, true)
@@ -311,11 +167,10 @@ class YourArticlesFragment : Fragment(R.layout.fragment_articles_list) {
                 is Resource.Error -> {
                     Util.isLoading(bounceLoader, false)
                     Util.isLoading(bounceLoaderBG, false)
-                    it.message?.let {
+                    selectedArticle.message?.let {
                         Log.e(TAG, "Error: $it")
                         Snackbar.make(requireView(), "Error: $it", Snackbar.LENGTH_SHORT).show()
                     }
-                    showYourArticles()
                 }
             }
         })
